+++ NPCs/Bosses/Greed/DesireSparkPro.cs
@@ -25,7 +25,7 @@ namespace AAMod.NPCs.Bosses.Greed
 
         public override void AI()
         {
-            projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
+            projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
             if (Main.rand.Next(4) == 0)
             {
                 int DustID2 = Dust.NewDust(projectile.position + projectile.velocity, projectile.width, projectile.height, 57, projectile.velocity.X * 0.2f, projectile.velocity.Y * 0.2f, 20, default(Color), 1f);
